Implemented the confirmation pop-up when deleting participants

+++ client/src/main/java/client/scenes/EditParticipantsCtrl.java
@@ -1,5 +1,6 @@
 package client.scenes;
 
+import client.components.Confirmation;
 import client.utils.LanguageConf;
 import client.utils.ServerUtils;
 import client.utils.Websocket;
@@ -8,11 +9,12 @@ import commons.Event;
 import commons.Participant;
 import javafx.beans.value.ObservableValue;
 import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.ChoiceBox;
-import javafx.scene.control.Label;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 import javafx.scene.text.Text;
+
+import java.util.Optional;
+
+import static java.lang.String.format;
 
 
 public class EditParticipantsCtrl {
@@ -155,12 +157,23 @@ public class EditParticipantsCtrl {
         mainCtrl.goBackToEventPage(event);
     }
 
+    /**
+     * When delete button is pressed, shows a confirmation screen
+     * and if confirmed deletes the participant
+     */
     @FXML
     private void deletePartClicked() {
         int index = chooseParticipant.getSelectionModel().getSelectedIndex();
         Participant part = event.getParticipants().get(index -1);
         String eventID = event.getId();
-        server.deleteParticipant(eventID, part.getId());
+        Confirmation confirmation =
+                new Confirmation((format(languageConf.get("EditP.deleteConfirmMessage"),
+                        part.getName())),
+                languageConf.get("Confirmation.areYouSure"), languageConf);
+        Optional<ButtonType> result = confirmation.showAndWait();
+        if(result.isPresent() && result.get() == ButtonType.YES) {
+            server.deleteParticipant(eventID, part.getId());
+        }
     }
 
     /**
